First attempt to create a stingray module.
This technique is probably not going to work. The problem is it's going to be very difficult to
make these methods declarative due to the imperative nature of the stingray cli and the
difficulty with executing arbitrary code in it and accessing output.

A better alternative might be to define a custom resource type where we can define the behaviours
more easily but this will take quite a white to get to a working state.

The faster approach is probably to simply copy files to the conf directory. This isn't documented
anywhere but it seems quite safe to do so.